mise en place de lombok

+++ src/main/java/com/jhilaa/tribean/jwt/JwtUtils.java
@@ -21,7 +21,7 @@ import java.util.stream.Collectors;
 @Component
 public class JwtUtils {
 
-    long JWT_VALIDITY = 5 * 60 * 60;
+    Integer JWT_VALIDITY = 5 * 60 * 60;
 
     private static final String AUTHORITIES_KEY = "sub";
